+++ frontend/src/interfaces/Bodypart.tsx
@@ -1,0 +1,12 @@
+const bodyPart = [
+	{ part: "Back", apiCall: "back" },
+	{ part: "Cardio", apiCall: "cardio" },
+	{ part: "Chest", apiCall: "chest" },
+	{ part: "Lower Arms", apiCall: "lower arms" },
+	{ part: "Lower Legs", apiCall: "lower legs" },
+	{ part: "Neck", apiCall: "neck" },
+	{ part: "Shoulders", apiCall: "shoulders" },
+	{ part: "Upper Arms", apiCall: "upper arms" },
+];
+
+export default bodyPart;
